requirements.txt
obs-websocket-py
subprocess
psutil
git+https://github.com/howprobable/py_helpers